ACT-3271: Separate new code into new file.
