+++ src/sdk/resources/policyLocationActions.ts
@@ -1,28 +1,7 @@
 import { IHttpClient } from '../http';
 import { getTypedResponse, BaseResource}  from './resource';
-import { PolicyLocationAction, PolicyLocationRule, PromptAttributes, TrackingGroupLocationActions, TrackingGroupLocationRules } from '../models';
+import { PolicyLocationAction, PromptAttributes, TrackingGroupLocationActions } from '../models';
 import { CancellationToken } from '../cancellation-token';
-
-// TODO: move and maybe rename (placeholders)?
-export class PolicyLocationRulesResource extends BaseResource {
-	constructor(httpClient: IHttpClient) {
-		super(httpClient);
-	}
-
-	private getResourcePath(policyId: string, locationId: string): string {
-		return `policies/${policyId}/locations/${locationId}/rule_placeholders`;
-	}
-
-	async getRules(policyId: string, locationId: string, params?: any, token?: CancellationToken): Promise<TrackingGroupLocationRules[]> {
-		const rules = await this.httpClient.get(this.getResourcePath(policyId, locationId), params, token);
-		return getTypedResponse<TrackingGroupLocationRules[]>(rules);
-	}
-
-	async update(policyId: string, locationId: string, body: PolicyLocationRule[], params?: any, token?: CancellationToken): Promise<boolean> {
-		const response = await this.httpClient.post(this.getResourcePath(policyId, locationId), body, params, token);
-		return response?.status === 200;
-	}
-}
 
 export class PolicyLocationActionsResource extends BaseResource {
 	constructor(httpClient: IHttpClient) {
